(all) updated regexp to escape '.' in chunkserver config (#569)
Normally I would've combined this with
https://github.com/moosefs/moosefs/pull/568 (caught them during the same
reading), but I think this one is possibly slightly more controversial?
:sweat_smile:

This string in the documentation looks like a regular expression, which
makes the unescaped `.` in it feel odd. I went for the "idiomatic Perl"
style `[.]` escaping because I personally find it more readable, but I'm
happy to swap to the more common `\.` or even just have the PR closed
instead because I don't feel strongly about it at all. :smile:

(I also double checked that the code wasn't actually using a regular
expression that needed to be updated, and indeed confirmed that it has
some code that does parsing and something like `c == '.'` for testing
this case and the regex is purely documentation. :+1:)